Move CvStream to utils.private.hpp
